Use comp to compose mj operations
Previously I was using the tf-> to compose transformations AND go between string and meta-joda. Now I'm just composing the mj transformations with "comp" or other standard Clojure fns, and wrapping with t-> to do the stringify/destringify when needed.
This seems more in line with what people would do if they were making a higher-level library. Mostly, the string versions will probably be for tests, storage, and import/export.