relation-bounded-interval
Introduced the concept of an interval defined (as it is in most other libraries) by its start and end, except here it is two other intervals, one with the :starts relationship and the other with the :ends relationship to the interval defined.

This is the first pass at this concept.