ISO-8601 string handling in threading macros
Improved the t-> (behavior changed) and introduced t->>. These are quite consistent with Clojure threading macros (code is slight modification of those).
Changed signature of some operations such as (enclosing :month) to be compatible with t->>  and more conventional.
Enhanced stringify/destring to handle sequences of times.
A few fixes and renames to fit these changes.