Switch to LocalDateTime in prep for timezones
Everything was done with Joda's DateTime object, but the class that corresponds most closely to the times up to this point is the LocalDateTime, which has no timezone information.
In preparation for actually incorporating TimeZones, the non-timezone nature of the current version is being made more explicit by using LocalTime.