Fixed stringify bug
Times with a scale of :day or above which have a timezone (a slightly awkward case, to be sure) don't have a clear-cut ISO format that I have found, so I made my own rule that there must be a T before the offset. This makes the dash character followed by a number unambiguous.
The T was being left out. Now it seems to be working.